This is CRS Quantum Documentation

The CRS Quantun package contains a set of useful functions for scientists using the Qube computer.

Shovel.shootuntil — Function

shootuntil(circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose=false, estimate=false)::shootuntilresult

Same as the other method except that no function is provided.

Example

julia> coeflin = [1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0]
julia> result = shootuntil(c1, 0.001, 0.05, coeflin, true)
julia> println(result)

shootuntil(fun::Function, circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose=false, estimate=false, ignorefun::Bool=false)::shootuntilresult

This is the generic function called by the two other methods

Shovel.shootuntilresult — Type

shootuntilresult is a structure containing the results of shootuntil function. It is intended to be reused by another package to perform analysis.

Members

γ::Float64

Δ::Float64

circuit::QuantumCircuit

samplesize::Int64

Proportions::Vector{Float64}

funvalue::Float64

variancefun::Float64

Shovel.shLaTeX — Function

shLaTeX(pattern::String, str::String)::String will substitute "pattern" inside str by "\pattern". This is mostly used when pattern = sin, cos or sqrt.

shLaTeX(str::String)::String will substitute "cos" by "\cos", "sin" by "\sin" and "sqrt" by "\sqrt" in str.

shLaTeX(c::QuantumCircuit, FName::String)::Bool

Will generate a file containing the LaTeX/quantikz code in the standalone documentclass.

Arguments

circuit::QuantumCircuit: a QuantumCircuit as defined by Snowflake

FName::String: the name of the file to create. Warning! It will overwrite if already existing. If omitted it will print to stdout

Example

julia> shLaTeX(circuit, "Foo.bar")

This is an example of the output file generated by ToLaTeX(). It can be copy&paste to any other Latex document.

\documentclass{standalone}
\usepackage{tikz}
\usetikzlibrary{quantikz}
\begin{document}

\begin{quantikz}
\lstick{q[1]: } &  \gate{H} &  \ctrl{1} &  \gate{X} &  \ctrl{1} &  \qw &  \qw &  \qw &  \ctrl{2} &  \gate{Z} &  \qw \\
\lstick{q[2]: } &  \qw &  \gate{X} &  \qw &  \gate{Z} &  \qw &  \qw &  \ctrl{2} & \qw &\qw & \qw \\
\lstick{q[3]: } &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \qw & \qw & \gate{Z} & \qw & \qw \\
\lstick{q[4]: } &  \qw &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \gate{Z} &  \qw &  \ctrl{-3} &  \qw 
\end{quantikz}
\end{document}

shLaTeX(mqc::shMQC, FName = "")::Bool

Arguments

mqc::MQC: a Meta Quantum Circuit as defined by Shovel

FName::String: the name of the file to create. Warning! It will overwrite if already existing. If omitted it will print to stdout

Example

julia> shLaTeX(mqc, "Foo.bar")

Shovel.shpush_gate! — Function

shpush_gate!(c::QuantumCircuit, gate::Gate, label::String)

Arguments

c::QuantumCircuit: a Snowflake QuantumCircuit.

gate::Gate: a Snowflake Gate.

label::String: a nice lable to supplement the standard Snowflake display_symbom in the QuantumCircuit pipeline

This allows to insert better labelling such as LaTeX code for improved clarity

Example

julia> shpush_gate!(c, rotation_z(1, sqrt(2.0*pi)), "R_z(\sqrt{2\pi})")
Quantum Circuit Object:
    id: 0197d410-788e-11ed-290c-adec58eae39f
    qubit_count: 1
    bit_count: 0
q[1]:--R_z(\sqrt{2\pi})--
    
julia> ToLaTeX(c)
\begin{quantikz}
\lstick{q[1]: } & \gate{R_z(\sqrt{2\pi}} & \qw
\end{quantikz}
true

The Quantikz/LaTeX file will display $R_z(\sqrt{2\pi})$ instead of "Rz(2.5066282746310002)".

Shovel.Plug — Type

Plug(c::QuantumCircuit, qb::Int)

Plug is a structure containing the UUID of a circuit and a qubit number. It is the basic element of a Connector. The only validation done is that the qubit number of the circuit is valid (>0 and <=qubit_count).

The operator "==" is defined for plugs.

Shovel.Connector — Type

Connector(plg1::Plug, plg2::Plug)
Connector(c1::QuantumCircuit, qb1::Int, c2::QuantumCircuit, qb2::Int)

Connector is a structure containing two plugs: 1) the input plug which is when the qubit/circuit is coming from and 2) the output plug indicating to which qubit/circuit it is going to. Users can either create plugs and then a connector from them or directly create a connector by providing the circuit and the qubit.

The operator "==" is defined for connector_list.

Shovel.isinverse — Function

isinverse(connec1::Connector, connec2::Connector)::Bool

A function to checks if a given connector is the inverse of another one. The function is used for internal consistency when a connector is added to an MQC. it will return true if "connec1.plugin == connec2.plugout && connec1.plugin == connec2.plugout" and false otherwise.

Shovel.isbefore — Function

isbefore(connec1::Connector, connec2::Connector)::Bool

A function to checks if the output plug of connec1 is the same as the input plug of connec2. If true, it means that connec1 is just before connec2 and they are connected together in the same wire.

Shovel.shprintlightQC — Function

shprintlightQC(io::IO, circuit::QuantumCircuit)

A quick display of basic info on a QuantumCircuit

Shovel.shMQC — Type

shMQC()

The structure shMQC is the main element of the Meta Quantum Circuit utility. After adding quantum circuits (or "circuits" for short) and connector_list, a quantikz/LaTeX file can be produced. Most importantly, a new circuit can be generated from the MQC.

Shovel.shMQCAddCircuit — Function

shMQCAddCircuit(mqc::MQC, newc::QuantumCircuit)::Bool

This function is used to add a Snowflake QuantumCircuit to an shMQC.

A given circuit cannot be add twice ot the shMQC. However, two distinct circuits with identical circuitry can.

The function will retrun true if the addition was successful.

Shovel.shMQCAddConnector — Function

shMQCAddConnector(mqc::shMQC, connec::Connector)::Bool

This function is used to add a connector to an MQC. It has some consistancy checks and will return false if the proposed connector creates inconsistencies such as circular circuitry or duplicate plugs.

Shovel.shsew — Function

shsew(mqc::shMQC)::QuantumCircuit

This function takes an MQC returns a standard Snowflake QuantumCircuit equivalent.

Shovel.printshMQC — Function

printshMQC(io::IO, mqc::shMQC)

Summary print of what's inside a shMQC

Shovel.shQuantumCircuit — Type

shQuantumCircuit is a structure holding additional descriptive information for rgw Quantum circuits such as a label for circuits and angle information for rotation gates Rx, Ry, R_y and the generic rotation gate.

Shovel.shoperator — Function

shoperator(c::QuantumCircuit)::Matrix{String} will produce the string version of the resulting operator of a quantum circuit

Shovel.shblock — Function

shblock(mat::Matrix{String}, bstart::Int, bsize::Int)::Matrix{String} will return the submatrix of a square matrix starting at position "bstart, bstart" and of size bsize x bsize