This is CRS Quantum Documentation

The CRS Quantun package contains a set of useful functions for scientists using the Qube computer.

Shovel.shootuntil — Function

function shootuntil(fun::Function, circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose=false, estimate=false)::shootuntilresult or function shootuntil(circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose=false, estimate=false)::shootuntilresult

Runs a circuit until there is a probability 1-γ that the precision Δ is reached for each of the state measurements.

Arguments

fun::function : is a function you want to calculate on the resulting proportion estimate on the final state of the circuit. For instance "sqrt" to get |α| instead of |α|^2

The function must take a Float64 as and input and return a Float64

circuit::QuantumCircuit: a QuantumCircuit as defined by Snowflake

Δ::Float64: the difference between the real value and the estimation

γ::Float64: the probability that the estimator is more that Δ apart from the true value.

linearcoef::Vector{Float64} : a vector of size 2^q, where q is the number of qubit in the circuit (q=circuit.qubit_count).`

It is a linear combination of the probabilities of the possible bit states after measurement. For more details please see here.

verbose::boolean: println usefull information on screen if needed for estimating suitable for Δ and γ.

estimate::boolean : this will prevent the fuction to run past the log(1-γ)/log(1-Δ) limit which is enough to get a rough estimation of the number of shots required to reach the desired precision.

Example

julia> coeflin = [1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0]
julia> result = shootuntil(c1, 0.001, 0.05, coeflin, true)
julia> println(result)

Shovel.ToLaTeX — Function

ToLaTeX(c::QuantumCircuit, FName::String)::Bool

Will generate a file containing the \LaTeX/quantikz code in the standalone documentclass.

Arguments

circuit::QuantumCircuit: a QuantumCircuit as defined by Snowflake

FName::String: the name of the file to create. Warning! It will overwrite if already existing.

Example

julia> ToLaTeX(circuit, "Foo.bar")

This is an example of the output file generated by ToLaTeX(). It can be copy&paste to any other Latex document.

\documentclass{standalone}
\usepackage{tikz}
\usetikzlibrary{quantikz}
\begin{document}

\begin{quantikz}
\lstick{q[1]: } &  \gate{H} &  \ctrl{1} &  \gate{X} &  \ctrl{1} &  \qw &  \qw &  \qw &  \ctrl{2} &  \gate{Z} &  \qw \\
\lstick{q[2]: } &  \qw &  \gate{X} &  \qw &  \gate{Z} &  \qw &  \qw &  \ctrl{2} & \qw &\qw & \qw \\
\lstick{q[3]: } &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \qw & \qw & \gate{Z} & \qw & \qw \\
\lstick{q[4]: } &  \qw &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \gate{Z} &  \qw &  \ctrl{-3} &  \qw 
\end{quantikz}
\end{document}

ToLaTeX(mqc::MQC, FName::String)::Bool

Will generate the \LaTeX/quantikz code representing the circuits and the connectors composing the Meta circuit.

Shovel.Plug — Type

Plug is a structure containing the UUID of a circuit and a qubit number. It is the basic element of a Connector. The only validation done is that the qubit number of the circuit is valid (>0 and <=qubit_count). The operator "==" is defined for plugs.

Shovel.Connector — Type

Connector is a structure containing two plugs: 1) the input plug which is when the qubit/circuit is coming from and 2) the output plug indicating to which qubit/circuit it is going to. Users can either create plugs and then a connector from them or directly create a connector by providing the circuit and the qubit. The operator "==" is defined for connector_list.

Shovel.isinverse — Function

isinverse(connec1::Connector, connec2::Connector)::Bool

A function to checks if a given connector is the inverse of another one. The function is used for internal consistency when a connector is added to an MQC. it will return true if "connec1.plugin == connec2.plugout && connec1.plugin == connec2.plugout" and false otherwise.

Shovel.isbefore — Function

isbefore(connec1::Connector, connec2::Connector)::Bool

A function to checks if the output plug of connec1 is the same as the input plug of connec2. if true, it means that connec1 is just before connec2 and they are connected together in the same wire.

Shovel.printlightQC — Function

printlightQC(io::IO, circuit::QuantumCircuit) A quick display of basic info on a QuantumCircuit

Shovel.MQC — Type

The structure MQC is the main element of the Meta Quantum Circuit utility. After adding quantum circuits (or "circuits" for short) and connector_list, a quantikz/LaTeX file can be produced. Most importantly, a new circuit can be generated from the MQC.

Shovel.MQCAddCircuit — Function

MQCAddCircuit(mqc::MQC, newc::QuantumCircuit)::Bool

This function is used to add a Snowflake QuantumCircuit to an MQC. A given circuit cannot be add twice ot the MQC. However, two distinct circuits with identical circuitry can. The function will retrun true if the addition was successful.

Shovel.MQCAddConnector — Function

MQCAddConnector(mqc::MQC, connec::Connector)::Bool

This function is used to add a connector to an MQC. It has some consistancy checks and will return false if the proposed connector creates inconsistencies such as circular circuitry or duplicate plugs.

Shovel.sew — Function

sew(mqc::MQC)

This function turns an MQC into a standard Snowflake QuantumCircuit

Shovel.printMQC — Function

printMQC(io::IO, mqc::MQC) Summary print of what's inside a MQC