Shovel.jl Documentation

The Shovel package contains a set of useful functions for scientists using the Qube computer.

Shovel.shootuntilresult — Type

shootuntilresult is a structure containing the results of shootuntil function. It is intended to be reused by another package to perform analysis.

Members

γ::Float64

Δ::Float64

circuit::QuantumCircuit

samplesize::Int64

Proportions::Vector{Float64}

funvalue::Float64

variancefun::Float64

Shovel.printshootresult — Function

printshootresult(io::IO, shrslt::shootuntilresult)

Pretty print of the shootuntilresult structure.

Shovel.shootuntil — Function

shootuntil(fun::Function, circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose::Bool=false, estimate::Bool=false, ignorefun::Bool=false)::shootuntilresult

Runs a circuit until there is a probability 1-γ that the precision Δ is reached for each of the state measurements.

Arguments

- `fun::function : is a function you want to calculate on the resulting proportion estimate on the final state of the circuit. For instance "sqrt" to get |α| instead of |α|^2
The function must take a Float64 as and input and return a Float64`
- `circuit::QuantumCircuit`: a QuantumCircuit as defined by Snowflake
- `Δ::Float64`: the difference between the real value and the estimation
- `γ::Float64`: the probability that the estimator is more that Δ apart from the true value.
- `linearcoef::Vector{Float64}`  : a vector of size 2^q, where q is the number of qubit in the circuit (q=circuit.qubit_count). It is a linear combination of the probabilities of the possible bit states after measurement. For more details please see [here](Stop/index.html).
- `verbose::boolean`: println usefull information on screen if needed for estimating suitable for Δ and γ. 
- `estimate::boolean` : this will prevent the fuction to run past the log(1-γ)/log(1-Δ) limit which is enough to get a rough estimation of the number of shots required 
to reach the desired precision.

Example

julia> coeflin = [1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0]
julia> result = shootuntil(sqrt, c1, 0.001, 0.05, coeflin, true)
julia> println(result)

shootuntil(circuit::QuantumCircuit, Δ::Float64, γ::Float64, linearcoef::Vector{Float64}, verbose=false, estimate=false)::shootuntilresult

Same as above but no function is provided.

Shovel.shLaTeX — Function

shLaTeX(pattern::String, str::String)::String

This function will substitute "pattern" inside str by "\pattern". This is mostly used when pattern = sin, cos or sqrt.

shLaTeX(str::String)::String

This function will substitute "cos" by "\cos", "sin" by "\sin" and "sqrt" by "\sqrt" in str.

shLaTeX(c::QuantumCircuit, FName::String)::Bool

Will generate a file containing the LaTeX/quantikz code in the standalone documentclass.

Arguments

circuit::QuantumCircuit: a QuantumCircuit as defined by Snowflake

FName::String: the name of the file to create. Warning! It will overwrite if already existing. If omitted it will print to stdout

Example

julia> shLaTeX(circuit, "Foo.bar")

This is an example of the output file generated by ToLaTeX(). It can be copy&paste to any other Latex document.

\documentclass{standalone}
\usepackage{tikz}
\usetikzlibrary{quantikz}
\begin{document}

\begin{quantikz}
\lstick{q[1]: } &  \gate{H} &  \ctrl{1} &  \gate{X} &  \ctrl{1} &  \qw &  \qw &  \qw &  \ctrl{2} &  \gate{Z} &  \qw \\
\lstick{q[2]: } &  \qw &  \gate{X} &  \qw &  \gate{Z} &  \qw &  \qw &  \ctrl{2} & \qw &\qw & \qw \\
\lstick{q[3]: } &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \qw & \qw & \gate{Z} & \qw & \qw \\
\lstick{q[4]: } &  \qw &  \qw &  \qw &  \qw &  \qw &  \gate{H} &  \gate{Z} &  \qw &  \ctrl{-3} &  \qw 
\end{quantikz}
\end{document}

shLaTeX(mat::Matrix{String})

Will convert to LaTeX the operator matrix generated by shoperator

shLaTeX(mqc::shMQC, FName = "")::Bool

Arguments

mqc::shMQC: a Meta Quantum Circuit as defined by Shovel

FName::String: the name of the file to create. Warning! It will overwrite if already existing. If omitted it will print to stdout

Example

julia> shLaTeX(mqc, "Foo.bar")

Shovel.shoperator — Function

shoperator(c::QuantumCircuit)::Matrix{String}

Returns, in string format, the resulting operator of a quantum circuit

Shovel.shinit1qubit! — Function

shinit1qubit!(probability::Float64)::QuantumCircuit

Will return a 1 qubit QuantumCircuit initialized such that its probability of being equal to 0 is equal to "probability"

Shovel.shinit2qubits! — Function

shinit2qubits!(probability::Vector{Float64})::QuantumCircuit

Will return a 2 qubits QuantumCircuit initialize such that the probabilities of the 4 possible outcome are equal to the values provided by the vector "probabilities"

Shovel.Plug — Type

Plug(c::QuantumCircuit, qb::Int)

Plug is a structure containing the UUID of a circuit and a qubit number. It is the basic element of a Connector. The only validation done is that the qubit number of the circuit is valid (>0 and <=qubit_count).

The comparison operator "==" is defined for plugs and returns true with both plugs connect the same qubits of the same circuits.

Shovel.Connector — Type

Connector(plg1::Plug, plg2::Plug)
Connector(c1::QuantumCircuit, qb1::Int, c2::QuantumCircuit, qb2::Int)

Connector is a structure containing two plugs: 1) the input plug which is when the qubit/circuit is coming from and 2) the output plug indicating to which qubit/circuit it is going to. Users can either create plugs and then a connector from them or directly create a connector by providing the circuit and the qubit.

The comparison operator "==" is defined for connectors and returns true if they have the same plugs from the same circuits IN THE SAME ORDER. If the two connectors have the same plugs but in REVERSE order, then function "isreverse" should be used to check.

Shovel.isinverse — Function

isinverse(connec1::Connector, connec2::Connector)::Bool

A function to checks if a given connector is the inverse of another one. The function is used for internal consistency when a connector is added to an MQC. it will return true if "connec1.plugin == connec2.plugout && connec1.plugin == connec2.plugout" and false otherwise.

Shovel.isbefore — Function

isbefore(connec1::Connector, connec2::Connector)::Bool

A function to checks if the output plug of connec1 is the same as the input plug of connec2. If true, it means that connec1 is just before connec2 and they are connected together in the same wire.

Shovel.shprintlightQC — Function

shprintlightQC(io::IO, circuit::QuantumCircuit)

A quick display of basic info on a QuantumCircuit

Shovel.shMQC — Type

shMQC

The structure shMQC is the main element of the Meta Quantum Circuit utility. After adding quantum circuits (or "circuits" for short) and connector_list, a quantikz/LaTeX file can be produced. Most importantly, a new circuit can be generated from the shMQC.

Shovel.printshMQC — Function

printshMQC(io::IO, mqc::shMQC)

Summary print of what's inside a shMQC

Shovel.shMQCAddCircuit — Function

shMQCAddCircuit(mqc::MQC, newc::QuantumCircuit)::Bool

This function is used to add a Snowflake QuantumCircuit to an shMQC. A given circuit cannot be add twice ot the shMQC. However, two distinct circuits with identical circuitry can as long as their id is different. The function will retrun true if the addition was successful.

Shovel.shMQCAddConnector — Function

shMQCAddConnector(mqc::shMQC, connec::Connector)::Bool

This function is used to add a connector to an MQC. It has some consistancy checks and will return false if the proposed connector creates inconsistencies such as circular circuitry or duplicate plugs.

Shovel.shsew — Function

shsew(mqc::shMQC)::QuantumCircuit

This function takes an shMQC and returns a standard Snowflake QuantumCircuit equivalent. This is the main goal of the shMQC concept.

Shovel.shQuantumCircuit — Type

shQuantumCircuit

shQuantumCircuit is a structure holding additional descriptive information for rgw Quantum circuits such as a label for circuits and angle information for rotation gates Rx, Ry, R_y and the generic rotation gate.